requirements.txt
aiogram==2.16
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.0.6
APScheduler==3.6.3
asgiref==3.5.0
async-generator==1.10
async-timeout==4.0.1
attrs==21.2.0
Babel==2.9.1
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
cachetools==4.2.2
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==36.0.0
distlib==0.3.3
Django==3.2.12
filelock==3.3.2
frozenlist==1.2.0
gitdb==4.0.9
GitPython==3.1.24
google-api-core==2.7.1
google-api-python-client==2.43.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-language==2.3.1
google-site-verification==0.0.4
googleapis-common-protos==1.53.0
grpcio==1.41.1
grpcio-status==1.41.1
gspread==5.0.0
gunicorn==20.1.0
h11==0.12.0
httplib2==0.20.2
idna==3.3
kombu==5.2.2
multidict==5.2.0
numpy==1.21.4
oauth2client==4.1.3
oauthlib==3.1.1
outcome==1.1.0
pandas==1.3.4
Pillow==8.4.0
platformdirs==2.4.0
prompt-toolkit==3.0.23
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.5
PySocks==1.7.1
pyTelegramBotAPI==4.2.0
python-dateutil==2.8.2
python-telegram-bot==13.8.1
python-wordpress-xmlrpc==2.3
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
schedule==1.1.0
selenium==4.1.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
tornado==6.1
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
virtualenv==20.10.0
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
wsproto==1.0.0
yarl==1.7.2
